OpenFragment/ContentFragment converted to ListFragment
ContentFragments now extend ListFragments, and the code for selection is
in place, if not yet working.  Code compiles and runs, but selection
causes FCs, and clicking on files produces no results.  Also, the
LogFragment has been disabled temporarily while it is converted to use
something other than the new ListFragmet as a base.