Big progress on Action Mode addition
Transition to ListView has been completed, and a new layout file is in
use for listview rows.  This layout uses a custom view which
encapsulates the OpenPath class, allowing OpenPaths to be associated
with the actual view.  This allows views to reference their associated
OpenPaths.  Checkboxes have also been added, although long presses are
the only current way to trigger actual selection.  More to come on that
front.  Currently FCs at setIcon() in ABS ActionBarImpl.java, so it
looks like the custom tab view needs to be tweaked before the Action
Mode will actually open.